backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
astropy==6.1.4
numpy==1.26.4
scipy==1.13.1
scikit-image==0.24.0
Pillow==10.4.0
pypng==0.20220715.0
colour-demosaicing==0.2.5
bm3d==4.0.3
pytest==8.3.3
